urs/backend/requirements.txt
asgiref==3.9.2
attrs==25.3.0
Django==5.0.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
